Bug 1560038 - Use i64 for storing epoch instead of usize. r=emilio
epoch is stored as double in ICU. Storing it as usize caused overflow on 32-bit Windows 7
which resulted in dates being formatted to epoch=0.
This patch switches to use f64 on the Rust side.

Depends on D66504

Differential Revision: https://phabricator.services.mozilla.com/D66872

+++ intl/l10n/rust/fluent-ffi/src/ffi.rs
@@ -146,7 +146,7 @@ extern "C" {
     ) -> *mut RawDateTimeFormatter;
     pub fn FluentBuiltInDateTimeFormatterFormat(
         formatter: *const RawDateTimeFormatter,
-        input: usize,
+        input: f64,
         out_count: &mut u32,
     ) -> *mut u8;
     pub fn FluentBuiltInDateTimeFormatterDestroy(formatter: *mut RawDateTimeFormatter);
